# perf: let EventedSet use clear() method of underlying set