--- a/benchmarks/benchmarks.py
+++ b/benchmarks/benchmarks.py
@@ -258,7 +258,7 @@ class EventedSetSuite:
     def time_update_overlap(self, n):
         self.my_set.update(range(n // 2, n + n // 2))
 
-    def time_clear(self):
+    def time_clear(self, _):
         self.my_set.clear()