--- a/src/psygnal/containers/_evented_set.py
+++ b/src/psygnal/containers/_evented_set.py
@@ -63,6 +63,12 @@ class _BaseMutableSet(MutableSet[_T]):
             self._do_discard(_item)
             self._post_discard_hook(_item)
 
+    def clear(self) -> None:
+        _item = self._pre_clear_hook()
+        if not isinstance(_item, BailType):
+            self._do_clear()
+            self._post_clear_hook(_item)
+
     def __contains__(self, value: object) -> bool:
         """Return True if value is in set."""
         return value in self._data
@@ -91,11 +97,19 @@ class _BaseMutableSet(MutableSet[_T]):
 
     def _post_discard_hook(self, item: _T) -> None: ...
 
+    def _pre_clear_hook(self) -> tuple[_T, ...] | BailType:
+        return tuple(self)  # pragma: no cover
+
+    def _post_clear_hook(self, item: tuple[_T, ...]) -> None: ...
+
     def _do_add(self, item: _T) -> None:
         self._data.add(item)
 
     def _do_discard(self, item: _T) -> None:
         self._data.discard(item)
+
+    def _do_clear(self) -> None:
+        self._data.clear()
 
     # -------- To match set API
 
@@ -299,6 +313,12 @@ class EventedSet(_BaseMutableSet[_T]):
 
     def _post_discard_hook(self, item: _T) -> None:
         self._emit_change((), (item,))
+
+    def _pre_clear_hook(self) -> tuple[_T, ...] | BailType:
+        return BAIL if len(self) == 0 else tuple(self)
+
+    def _post_clear_hook(self, item: tuple[_T, ...]) -> None:
+        self._emit_change((), item)
 
     def _emit_change(self, added: tuple[_T, ...], removed: tuple[_T, ...]) -> None:
         """Emit a change event."""