@tlambert03 2024-05-06
wow.  thanks this is fantastic.  LGTM

Sorry about the weird `BAIL` stuff... had to remind myself why it's there (clearly not obvious), I think it was mostly there for things like `add` where subclasses might want the ability to cancel the event, or to avoid an event entirely as you've done in the case if `len(self) == 0`.   but your PR is perfect, thanks!