@tlambert03 2024-05-06
btw, the codspeed benchmarks shown above don't capture this, but the asv benchmarks do and are awesome 🤩 


| Change   | Before [b0c29526] <main>   | After [3b7af9d8]    |   Ratio | Benchmark (Parameter)                                     |
|----------|----------------------------|---------------------|---------|-----------------------------------------------------------|
| -        | 1.15±0.01μs                | 689±7ns             |    0.6  | benchmarks.EventedSetWithCallbackSuite.time_clear(10)     |
| -        | 1.16±0.03μs                | 682±2ns             |    0.59 | benchmarks.EventedSetSuite.time_clear(10)                 |
| -        | 4.83±0s                    | 763±2ns             |    0    | benchmarks.EventedSetSuite.time_clear(100000)             |
| -        | 4.84±0.02s                 | 794±7ns             |    0    | benchmarks.EventedSetWithCallbackSuite.time_clear(100000) |

"infinite speedup" :joy: